refactor : service 클래스 interface로 추상화
